fix(app): address runtime review findings
